move validateMainCredentialData out of the shared module
Only the getOne test is using it.
