+++ packages/cli/test/integration/credentials/shared.ts
@@ -1,5 +1,3 @@
-import type { ListQuery } from '@/requests';
-
 import { randomName, randomString } from '../shared/random';
 
 export const INVALID_PAYLOADS = [
@@ -26,26 +24,3 @@ export const INVALID_PAYLOADS = [
 	{},
 	undefined,
 ];
-
-export function validateMainCredentialData(
-	credential: ListQuery.Credentials.WithOwnedByAndSharedWith,
-) {
-	const { name, type, nodesAccess, sharedWith, ownedBy } = credential;
-
-	expect(typeof name).toBe('string');
-	expect(typeof type).toBe('string');
-	expect(typeof nodesAccess?.[0].nodeType).toBe('string');
-
-	if (sharedWith) {
-		expect(Array.isArray(sharedWith)).toBe(true);
-	}
-
-	if (ownedBy) {
-		const { id, email, firstName, lastName } = ownedBy;
-
-		expect(typeof id).toBe('string');
-		expect(typeof email).toBe('string');
-		expect(typeof firstName).toBe('string');
-		expect(typeof lastName).toBe('string');
-	}
-}
